Extract duplicate text helper into a shared scope
