+++ EssentailFeed/EssentailFeedTests/Feed Cache/ValidateFeedCacheUseCaseTests.swift
@@ -55,33 +55,4 @@ class ValidateFeedCacheUseCaseTests: XCTestCase {
         return (sut, store)
     }
     
-    
-    private func anyNSError() -> NSError {
-        return NSError(domain: "any Error", code: 0)
-    }
-    
-    private func anyURL() -> URL{
-       return URL(string: "http://any-url.com")!
-    }
-    
-    private func uniqueImage() -> FeedImage {
-        return FeedImage(id: UUID(), description: "any desc", location: "any loc", url: anyURL())
-    }
-    
-    private func uniqueImageFeed() -> (modules: [FeedImage], local: [LocalFeedImage]) {
-        let modules = [uniqueImage(),uniqueImage()]
-        let local = modules.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url) }
-        
-        return (modules, local)
-    }
-}
-
-private extension Date {
-    func adding(days: Int) -> Date {
-        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding(seconds: TimeInterval) -> Date {
-        return self + seconds
-    }
 }
